+++ src/main/java/com/semicolon/dto/request/ItemDtoRequest.java
@@ -11,5 +11,5 @@ public class ItemDtoRequest {
     private String itemId;
     private Product product;
     private int quantity;
-    private BigDecimal subTotal;
+    private double subTotal;
 }
