+++ src/main/java/com/semicolon/dto/request/ProductDtoRequest.java
@@ -13,5 +13,5 @@ public class ProductDtoRequest {
     private String productName;
     private String productDescription;
     private ProductCategory productCategory;
-    private BigDecimal productPrice;
+    private double productPrice;
 }
